STT-NLP/requirements.txt
sounddevice
numpy
webrtcvad-wheels
scipy
torch
torchaudio
librosa
tqdm
pandas
PyYAML
matplotlib
seaborn
tinydb
spacy==3.7.4
streamlit
mlflow
